Add NotesContainer and Note component
A note component has functionality with states so when users type,
states are updated based on input. Each note has a title p element and a
content p element as well as inputs that correspond with each. When the
user enters their input, the state is updated and the p elements change
too. This is so the input will disappear when users are done editing
and will be replaced by p elements.