Add function to auto-focus on visible inputs
When user clicks a p element for either the note's title or content, the
input will autofocus so users can start typing right away. This is also
great because if a user isn't focus on an input it will always be a p
element, with the initial note content input as an exception.